Alternative Tiles can now be added!
Can now add alternative tiles to each bitmask and  set a chance they appear. Default tile (tile 0) will always be 100% (fallback).

+++ addons/threaded_autotiler/Scripts/BitmaskTile.cs
@@ -7,9 +7,9 @@ public class BitmaskTile
 
     public Vector2I AtlasValue;
 
-    public double Chance;
+    public float Chance;
 
-    public BitmaskTile(Vector2I atlasValue, ushort bitmaskValue, double chance = 100)
+    public BitmaskTile(Vector2I atlasValue, ushort bitmaskValue, float chance = 100)
     {
         AtlasValue = atlasValue;
         BitmaskValue = bitmaskValue;
